requirements.txt
# Build System Dependencies
jinja2>=3.1.0
pyyaml>=6.0.0
markdown>=3.5.0

# AI Image Manager Dependencies (Phase 2)
openai>=1.3.0
Pillow==10.0.1
PyYAML==6.0.1
python-dotenv==1.0.0
Markdown==3.5.1
requests==2.31.0
# Quality Gates & Validation (Phase 3)
html5lib>=1.1
beautifulsoup4>=4.12.0
cssutils>=2.9.0
